+++ framework/packages/abstract-core/src/core/manager.rs
@@ -21,7 +21,7 @@ pub mod state {
     use crate::objects::common_namespace::OWNERSHIP_STORAGE_KEY;
     use crate::objects::module::ModuleInfo;
     use crate::objects::{gov_type::GovernanceDetails, module::ModuleId};
-    use cosmwasm_std::{Addr, Api, Binary};
+    use cosmwasm_std::{Addr, Binary, Deps};
     use cw_address_like::AddressLike;
     use cw_controllers::Admin;
     use cw_ownable::Ownership;
@@ -48,8 +48,12 @@ pub mod state {
 
     impl AccountInfo<String> {
         /// Check an account's info, verifying the gov details.
-        pub fn verify(self, api: &dyn Api) -> Result<AccountInfo<Addr>, crate::AbstractError> {
-            let governance_details = self.governance_details.verify(api)?;
+        pub fn verify(
+            self,
+            deps: Deps,
+            version_control_addr: Addr,
+        ) -> Result<AccountInfo<Addr>, crate::AbstractError> {
+            let governance_details = self.governance_details.verify(deps, version_control_addr)?;
             Ok(AccountInfo {
                 name: self.name,
                 governance_details,
@@ -89,6 +93,10 @@ pub mod state {
     pub const DEPENDENTS: Map<ModuleId, HashSet<String>> = Map::new("dependents");
     /// Stores a queue of modules to install on the account after creation.
     pub const MODULE_QUEUE: Item<Vec<(ModuleInfo, Option<Binary>)>> = Item::new("mqueue");
+    /// List of sub-accounts
+    pub const SUB_ACCOUNTS: Map<u32, cosmwasm_std::Empty> = Map::new("sub_accs");
+    /// Pending new governance
+    pub const PENDING_GOVERNANCE: Item<GovernanceDetails<Addr>> = Item::new("pgov");
 }
 
 use self::state::AccountInfo;
@@ -137,6 +145,19 @@ pub enum InternalConfigAction {
     },
 }
 
+#[cosmwasm_schema::cw_serde]
+#[non_exhaustive]
+pub enum UpdateSubAccountAction {
+    /// Unregister sub-account
+    /// It will unregister sub-account from the state
+    /// Could be called only by the sub-account itself
+    UnregisterSubAccount { id: u32 },
+    /// Register sub-account
+    /// It will register new sub-account into the state
+    /// Could be called by the sub-account manager
+    /// Note: since it happens after the claim by this manager state won't have spam accounts
+    RegisterSubAccount { id: u32 },
+}
 /// Manager execute messages
 #[cw_ownable::cw_ownable_execute]
 #[cosmwasm_schema::cw_serde]
@@ -187,12 +208,14 @@ pub enum ExecuteMsg {
         link: Option<String>,
     },
     /// Sets a new Owner
-    /// New owner will have to claim ownership, in case force is not true
+    /// New owner will have to claim ownership
     SetOwner { owner: GovernanceDetails<String> },
     /// Update account statuses
     UpdateStatus { is_suspended: Option<bool> },
     /// Update settings for the Account, including IBC enabled, etc.
     UpdateSettings { ibc_enabled: Option<bool> },
+    /// Actions called by internal or external sub-accounts
+    UpdateSubAccount(UpdateSubAccountAction),
     /// Callback endpoint
     Callback(CallbackMsg),
 }
@@ -226,6 +249,11 @@ pub enum QueryMsg {
     /// Returns [`InfoResponse`]
     #[returns(InfoResponse)]
     Info {},
+    #[returns(SubAccountIdsResponse)]
+    SubAccountIds {
+        start_after: Option<u32>,
+        limit: Option<u8>,
+    },
 }
 
 #[cosmwasm_schema::cw_serde]
@@ -261,4 +289,9 @@ pub struct ManagerModuleInfo {
 #[cosmwasm_schema::cw_serde]
 pub struct ModuleInfosResponse {
     pub module_infos: Vec<ManagerModuleInfo>,
+}
+
+#[cosmwasm_schema::cw_serde]
+pub struct SubAccountIdsResponse {
+    pub sub_accounts: Vec<u32>,
 }